Fixed horrible bug in wrapped PApplet variables.
Because, for example, mouseX always referred to the same object, if you were
to add it to a list, then that list element would always refer to the current
mouseX, and not to the mouseX at the time it was added.

So we have to create new wrapped variables for every setup() and draw(), which
is scary for the amount of garbage it generates.