Fix re-initialization show-stopper.
My attempt to re-initialize the Jython state was leading to a really
vexing situation in which Py.None (amongst many other things) was
set to a new value, but some static somewhere had a reference to the
original Py.None, so that a None value, when == compared with Py.None,
was not equal! This made the csv module throw a completely inscrutable
error at import time.

Even worse, this conflated with the very simple problen that led to
csv not being importable in the first place (which was due to having
switched to an unzipped Jython distribution, without remembering to
tell Jython where its Lib was).

It has been a rough few hours.