Add libraries to classloader dynamically at runtime.
I consider this a necessary but sad design regression. In playing with the
beads library, I discovered that SPI lookups (for mp3 audio file format,
for example) were failing, because adding the jar files to the Jython
interpreter path was not enough for the entirely java-world framework
lookups to succeed. So now we have what I think it s brittle hack, to call
a protected method on the URLClassLoader to add library jar files to
the classpath.