Allowing running via a symbolic link
Check if the script is being accessed via a symbolic link, and if so correctly detect the actual location of the script so that Processing-py jar file and libraries will be accessible.

I install the package under /opt on an OS X machine. The extra symlink-resolution step is required for using a symlink from /usr/local/bin to the actual script's location.