Bring IDE to front when a sketch finishes on OSX
The IDE runs in a java environment whos java.library.path doesn't
include the directory containing jniosx, so the library must be loaded
using a path to the file. Overloaded OSX.bringToFront to allow using a
PythonMode object to create the path and load jniosx so we can bring the
IDE to front.

Removed the static initializer that loaded jniosx and moved the
System.loadLibrary command inside bringToFront so the two different
methods of loading the library wouldn't step on each other.

Added brinToFront(mode) to PyEditor to bring the IDE to
front.