Fix add_library for windows. Fix #308. Fix #227.
Adding paths to java.library.path in order to load a .dll wasn't working
on windows. I'm not sure how this is working on any platform, but
everything seems to be working fine. Split recursivelyAddToClasspath
into two different methods, one to add .jars to the classpath, and one
to load .dlls using the absolute file path.

These changes are only for the windows platform, all other platforms
function the old way.